datacollection/requirements.txt
django
django-celery
redis
rpyc
